Add files via upload
In this project, program for the SMS Spam Detection based on character n-gram 
models approach is explained. In order to implement the program, jupyter 
notebook is used. For training and testing the model the program used UCI 
Machine Learning Repository SMS Spam Collection dataset. In order to use the 
dataset for n-gram model pre-processing is performed to clean the dataset. The 
dataset is then divided into training and test sets by taking 80% of dataset as 
training dataset and remaining as test dataset.